+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Packaging module to nupkg
 - for new features
 
 ### Changed
@@ -30,3 +31,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - for now removed features.
+
+## [v0.0.1] 2019-10-04
+
+### Added
+
+- dummy release for example